refactor(rust): lib.rsのコマンドをドメイン別モジュールに分割
35個のTauriコマンドをcommands/{library,work,tag,settings,import}.rs
に分割し、lib.rsはAppState/AppDb/ActiveLibraryの定義とrun()のみに縮小。
migrate_libraries_jsonはmigration.rsへ移設。generate_handler!はフルパス
参照に更新。
