+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod import;
+pub(crate) mod library;
+pub(crate) mod settings;
+pub(crate) mod tag;
+pub(crate) mod work;
